+++ gulpfile.js
@@ -21,13 +21,13 @@ var gulp = require('gulp'),
             'src/component/view/DOMViewProvider.js',
             'src/component/view/ViewManager.js',
     		'src/directive/Directive.js',
-            'src/converter/Converter.js',
+            'src/filter/Filter.js',
             'src/service/Service.js',
 
             'src/factory/Factory.js',
             'src/factory/ComponentFactory.js',
             'src/factory/DirectiveFactory.js',
-            'src/factory/ConverterFactory.js',
+            'src/factory/FilterFactory.js',
             'src/factory/ServiceFactory.js',
                         
     		'src/impex.js',
@@ -36,8 +36,6 @@ var gulp = require('gulp'),
 
     		'src/service/build-in.js',
             'src/directive/build-in.js',
-            'src/filter/build-in.js',
-            'src/converter/build-in.js',
 
     		'src/shellEnd.js'
     	])